Moves CommunityHeader component to ui folder
