+++ pages/experiments/index.vue
@@ -36,7 +36,7 @@
 <script lang="ts">
 import { Component } from 'vue-property-decorator'
 import QiskitPage from '~/components/qiskit/QiskitPage.vue'
-import CommunityHeader from '~/components/headers/CommunityHeader.vue'
+import CommunityHeader from '~/components/ui/CommunityHeader.vue'
 import PageSection from '~/components/sections/PageSection.vue'
 import ExperimentCard from '~/components/experiments/ExperimentCard.vue'
 import ExperimentCarousel from '~/components/experiments/ExperimentCarousel.vue'
